pyproject.toml
[tool.poetry]
name = "postdatabase"
version = "0.1.0"
description = ""
authors = ["gabriel.mango <gabriel.mango@defensoria.mg.def.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
